refactor: extract detect-stack into shared lib/ for cross-skill reuse
Move stack detection logic (STACK_MARKERS, STACK_META, detectStack,
gitignoreTemplates, dependabotEcosystems) from repo-ready/scripts/ into
lib/detect-stack.mjs as the single source of truth.

The skill-local scripts/detect-stack.mjs becomes a thin re-export plus
CLI entry point. All existing imports, tests, and CLI usage continue
working unchanged.

This establishes the lib/ convention: shared modules live at repo root,
skills re-export them locally for self-contained imports and CLI.

Copilot-Session: fa6ffb1e-188e-4046-abd0-b4fb8868963e
